refactor ffmpeg specfic code to ffmpeg util
